+++ src/options/components/single-processor-tab.tsx
@@ -228,7 +228,11 @@ export function SingleProcessorTab() {
             ? {
                 tinyMode: formatOptions.png.tinyMode,
                 cleanTransparentPixels: formatOptions.png.cleanTransparentPixels,
-                autoGrayscale: formatOptions.png.autoGrayscale
+                autoGrayscale: formatOptions.png.autoGrayscale,
+                dithering: formatOptions.png.dithering,
+                ditheringLevel: formatOptions.png.ditheringLevel,
+                progressiveInterlaced: formatOptions.png.progressiveInterlaced,
+                oxipngCompression: formatOptions.png.oxipngCompression
               }
             : undefined
       },